Exclude policy logic for linux
